Fix Invalid Argument in chmod if mode=-1 (fixes 2033)
git-svn-id: svn://svn.lighttpd.net/spawn-fcgi/trunk@52 4a9f3682-ca7b-49a8-9a55-ba4640e46f83

+++ src/spawn-fcgi.c
@@ -200,7 +200,7 @@ static int bind_socket(const char *addr, unsigned short port, const char *unixso
 			}
 		}
 
-		if (-1 == chmod(unixsocket, mode)) {
+		if (-1 != mode && -1 == chmod(unixsocket, mode)) {
 			fprintf(stderr, "spawn-fcgi: couldn't chmod socket: %s\n", strerror(errno));
 			close(fcgi_fd);
 			unlink(unixsocket);
